+++ pytorch_lightning/plugins/training_type/horovod.py
@@ -34,7 +34,7 @@ if _HOROVOD_AVAILABLE:
     import horovod.torch as hvd
 
 
-class HorovodPlugin(ParallelStrategy):
+class HorovodStrategy(ParallelStrategy):
     """Plugin for Horovod distributed training integration."""
 
     distributed_backend = _StrategyType.HOROVOD